Move geonode to src/gui.
