+++ src/gui/geonode/qgsgeonodenewconnection.cpp
