+++ src/gui/geonode/qgsgeonodesourceselect.cpp
